Added UserController and ProjectController Ingtegration tests

+++ src/main/java/com/bluesky/bugtraker/service/impl/UserServiceImp.java
@@ -12,7 +12,8 @@ import com.bluesky.bugtraker.io.repository.TicketRepository;
 import com.bluesky.bugtraker.io.repository.UserRepository;
 import com.bluesky.bugtraker.security.UserPrincipal;
 import com.bluesky.bugtraker.service.UserService;
-import com.bluesky.bugtraker.service.utils.ServiceUtils;
+import com.bluesky.bugtraker.service.utils.DataExtractionUtils;
+import com.bluesky.bugtraker.service.utils.UserServiceUtils;
 import com.bluesky.bugtraker.service.utils.Utils;
 import com.bluesky.bugtraker.shared.authorizationenum.Role;
 import com.bluesky.bugtraker.shared.dto.ProjectsInfoDTO;
@@ -34,10 +35,12 @@ import org.springframework.transaction.annotation.Transactional;
 import javax.validation.constraints.NotNull;
 import java.time.Instant;
 import java.time.Period;
-import java.util.*;
+import java.util.Date;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
-import static com.bluesky.bugtraker.exceptions.ErrorType.NO_RECORD_FOUND;
-import static com.bluesky.bugtraker.exceptions.ErrorType.RECORD_ALREADY_EXISTS;
+import static com.bluesky.bugtraker.exceptions.ErrorType.*;
 import static com.bluesky.bugtraker.service.specifications.Specs.*;
 
 @Service
@@ -45,29 +48,25 @@ public class UserServiceImp implements UserService {
     @Value("${user-avatar-url}")
     private String avatarURL;
 
-    private final ServiceUtils serviceUtils;
     private final UserRepository userRepo;
     private final RoleRepository roleRepo;
     private final TicketRepository ticketRepo;
     private final ProjectRepository projectRepo;
 
 
+    private final DataExtractionUtils dataExtractionUtils;
+    private final UserServiceUtils userServiceUtils;
     private final Utils utils;
     private final ModelMapper modelMapper;
 
     @Autowired
-    public UserServiceImp(ServiceUtils serviceUtils,
-                          UserRepository userRepo,
-                          RoleRepository roleRepo,
-                          TicketRepository ticketRepo,
-                          ProjectRepository projectRepo,
-                          Utils utils,
-                          ModelMapper modelMapper) {
-        this.serviceUtils = serviceUtils;
+    public UserServiceImp(DataExtractionUtils dataExtractionUtils, UserRepository userRepo, RoleRepository roleRepo, TicketRepository ticketRepo, ProjectRepository projectRepo, UserServiceUtils userServiceUtils, Utils utils, ModelMapper modelMapper) {
+        this.dataExtractionUtils = dataExtractionUtils;
         this.userRepo = userRepo;
         this.roleRepo = roleRepo;
         this.ticketRepo = ticketRepo;
         this.projectRepo = projectRepo;
+        this.userServiceUtils = userServiceUtils;
 
         this.utils = utils;
         this.modelMapper = modelMapper;
@@ -89,7 +88,7 @@ public class UserServiceImp implements UserService {
     @Override
     @NotNull
     public UserDTO getUserById(@NotNull String id) {
-        UserEntity userEntity = serviceUtils.getUserEntity(id);
+        UserEntity userEntity = dataExtractionUtils.getUserEntity(id);
         UserDTO userDTO = modelMapper.map(userEntity, UserDTO.class);
         userDTO.setRole(userEntity.getRoleEntity().getRole());
 
@@ -99,8 +98,7 @@ public class UserServiceImp implements UserService {
     @Override
     @NotNull
     public UserDTO getUserByEmail(@NotNull String email) {
-        UserEntity userEntity = userRepo.findByEmail(email)
-                .orElseThrow(() -> new UserServiceException(NO_RECORD_FOUND, email));
+        UserEntity userEntity = userRepo.findByEmail(email).orElseThrow(() -> new UserServiceException(NO_RECORD_FOUND, email));
 
         UserDTO userDTO = modelMapper.map(userEntity, UserDTO.class);
         userDTO.setRole(userEntity.getRoleEntity().getRole());
@@ -110,7 +108,7 @@ public class UserServiceImp implements UserService {
 
     @Override
     @NotNull
-    public DataTablesOutput<UserDTO> getUsers(DataTablesInput input) {
+    public DataTablesOutput<UserDTO> getUsers(@NotNull DataTablesInput input) {
 
         DataTablesOutput<UserEntity> userEntities = userRepo.findAll(input);
 
@@ -120,6 +118,7 @@ public class UserServiceImp implements UserService {
 
 
     @Transactional
+    @NotNull
     public UserDTO createUserWithRole(@NotNull UserDTO userDto) {
         if (userRepo.existsByEmail(userDto.getEmail()))
             throw new UserServiceException(RECORD_ALREADY_EXISTS, userDto.getEmail());
@@ -131,9 +130,7 @@ public class UserServiceImp implements UserService {
         userEntity.setEmailVerificationToken("mock");
         userEntity.setEmailVerificationStatus(true);   // TODO after testing change to false;
 
-        RoleEntity roleEntity = roleRepo.findByRole(userDto.getRole())
-                .orElseThrow(() ->
-                        new RoleServiceException(NO_RECORD_FOUND, userDto.getRole().toString()));
+        RoleEntity roleEntity = roleRepo.findByRole(userDto.getRole()).orElseThrow(() -> new RoleServiceException(NO_RECORD_FOUND, userDto.getRole().toString()));
         userEntity.setRoleEntity(roleEntity);
 
         UserEntity savedEntity = userRepo.save(userEntity);
@@ -142,24 +139,23 @@ public class UserServiceImp implements UserService {
     }
 
     @Override
-    public UserDTO createUser(UserDTO userDto) {
+    @NotNull
+    public UserDTO createUser(@NotNull UserDTO userDto) {
         userDto.setRole(Role.ROLE_USER);
         return createUserWithRole(userDto);
     }
 
     @Override
-    public void updateUser(String userId, UserDTO userDto) {
-        UserEntity userEntity = serviceUtils.getUserEntity(userId);
+    public void updateUser(@NotNull String userId, @NotNull UserDTO userDto) {
+        UserEntity userEntity = dataExtractionUtils.getUserEntity(userId);
 
         if (userDto.getRole() != null) {
-            RoleEntity roleEntity =
-                    serviceUtils.getRoleEntityToBeSet(userDto.getRole(), userEntity.getRoleEntity().getRole());
+            RoleEntity roleEntity = dataExtractionUtils.getRoleEntityToBeSet(userDto.getRole(), userEntity.getRoleEntity().getRole());
             userEntity.setRoleEntity(roleEntity);
         }
 
         String avatarURL = userDto.getAvatarURL();
-        if (avatarURL == null || avatarURL.isBlank())
-            userDto.setAvatarURL(this.avatarURL);
+        if (avatarURL == null || avatarURL.isBlank()) userDto.setAvatarURL(this.avatarURL);
 
         modelMapper.map(userDto, userEntity);
         userRepo.save(userEntity);
@@ -167,62 +163,55 @@ public class UserServiceImp implements UserService {
 
     @Override
     public void deleteUser(@NotNull String userId) {
-        UserEntity userEntity = serviceUtils.getUserEntity(userId);
+        UserEntity userEntity = dataExtractionUtils.getUserEntity(userId);
+
+        if(userEntity.getRoleEntity().getRole() == Role.ROLE_SUPER_ADMIN
+                && userRepo.countByRoleEntityRole(Role.ROLE_SUPER_ADMIN) <= 1)
+            
+            throw new UserServiceException(INTERNAL_SERVER_ERROR,
+                    "At least one user with role " + Role.ROLE_SUPER_ADMIN + " should exists");
+
         userRepo.delete(userEntity);
     }
 
     @Override
     @NotNull
     public ProjectsInfoDTO getProjectsInfo(String userId) {
-        UserEntity userEntity = serviceUtils.getUserEntity(userId);
+        UserEntity userEntity = dataExtractionUtils.getUserEntity(userId);
 
-        Date dayAgo =  Date.from(Instant.now().minus(Period.ofDays(1)));
+        Date dayAgo = Date.from(Instant.now().minus(Period.ofDays(1)));
         ProjectsInfoDTO projectsInfoDTO = new ProjectsInfoDTO();
 
-        projectsInfoDTO.setTicketsReported(
-                ticketRepo.countAllByProjectCreator(userEntity));
-        projectsInfoDTO.setTicketsReportedADayAgo(
-                ticketRepo.countAllByProjectCreatorAndLastUpdateTimeAfter(userEntity, dayAgo));
+        long ticketsReported = ticketRepo.countAllByProjectCreator(userEntity);
+        long ticketsReportedADayAgo = ticketRepo.countAllByProjectCreatorAndLastUpdateTimeAfter(userEntity, dayAgo);
+        
+        long criticalTickets = ticketRepo.countAllByProjectCreatorAndSeverity(userEntity, Severity.CRITICAL);
+        long criticalTicketsADayAgo = ticketRepo.countAllByProjectCreatorAndSeverityAndLastUpdateTimeAfter(userEntity, Severity.CRITICAL, dayAgo);
+        
+        long completedTickets = ticketRepo.countAllByProjectCreatorAndStatus(userEntity, Status.COMPLETED);
+        long completedTicketsADayAgo = ticketRepo.countAllByProjectCreatorAndStatusAndLastUpdateTime(userEntity, Status.COMPLETED, dayAgo);
+       
+        long ticketsInProgress = ticketRepo.countAllByProjectCreatorAndStatus(userEntity, Status.IN_PROGRESS);
+        long ticketInProgressADayAgo = ticketRepo.countAllByProjectCreatorAndStatusAndLastUpdateTime(userEntity, Status.IN_PROGRESS, dayAgo);
+       
+        projectsInfoDTO.setTicketsReported(ticketsReported);
+        projectsInfoDTO.setTicketsReportedADayAgo(ticketsReportedADayAgo);
 
-        projectsInfoDTO.setCriticalTickets(
-                ticketRepo.countAllByProjectCreatorAndSeverity(userEntity, Severity.CRITICAL));
-        projectsInfoDTO.setCriticalTicketsADayAgo(
-                ticketRepo.countAllByProjectCreatorAndSeverityAndLastUpdateTimeAfter(userEntity, Severity.CRITICAL, dayAgo));
+        projectsInfoDTO.setCriticalTickets(criticalTickets);
+        projectsInfoDTO.setCriticalTicketsADayAgo(criticalTicketsADayAgo);
 
-        projectsInfoDTO.setCompletedTickets(
-                ticketRepo.countAllByProjectCreatorAndStatus(userEntity, Status.COMPLETED));
-        projectsInfoDTO.setCompletedTicketsADayAgo(
-                ticketRepo.countAllByProjectCreatorAndStatusAndLastUpdateTime(userEntity, Status.COMPLETED, dayAgo));
+        projectsInfoDTO.setCompletedTickets(completedTickets);
+        projectsInfoDTO.setCompletedTicketsADayAgo(completedTicketsADayAgo);
 
-        projectsInfoDTO.setTicketsInProgress(
-                ticketRepo.countAllByProjectCreatorAndStatus(userEntity, Status.IN_PROGRESS));
-        projectsInfoDTO.setTicketInProgressADayAgo(
-                ticketRepo.countAllByProjectCreatorAndStatusAndLastUpdateTime(userEntity, Status.IN_PROGRESS, dayAgo));
+        projectsInfoDTO.setTicketsInProgress(ticketsInProgress);
+        projectsInfoDTO.setTicketInProgressADayAgo(ticketInProgressADayAgo);
 
 
-        Map<Status, List<Long>> ticketsPerWeek = new EnumMap<>(Status.class);
-        projectsInfoDTO.setTicketsPerWeek(ticketsPerWeek);
+        Map<Status, List<Long>> ticketsStatusPerWeek = userServiceUtils.getTicketsStatusPerWeek(userEntity);
+        projectsInfoDTO.setTicketsStatusPerWeek(ticketsStatusPerWeek);
 
-        for (int i = 0; i < 5; i++) {
-            long countToFix = ticketRepo.count(ticketByProjectCreator(userEntity).and(statusIs(Status.TO_FIX).and(lastUpdatedWeeksAgo(i))));
-            long countInProgress = ticketRepo.count(ticketByProjectCreator(userEntity).and(statusIs(Status.IN_PROGRESS).and(lastUpdatedWeeksAgo(i))));
-            long countCompleted = ticketRepo.count(ticketByProjectCreator(userEntity).and(statusIs(Status.COMPLETED).and(lastUpdatedWeeksAgo(i))));
-
-            ticketsPerWeek.putIfAbsent(Status.TO_FIX, new ArrayList<>());
-            ticketsPerWeek.putIfAbsent(Status.IN_PROGRESS, new ArrayList<>());
-            ticketsPerWeek.putIfAbsent(Status.COMPLETED, new ArrayList<>());
-
-            ticketsPerWeek.get(Status.TO_FIX).add(countToFix);
-            ticketsPerWeek.get(Status.IN_PROGRESS).add(countInProgress);
-            ticketsPerWeek.get(Status.COMPLETED).add(countCompleted);
-        }
-
-        Map<Status, Long> tickets = new EnumMap<>(Status.class);
-        projectsInfoDTO.setTickets(tickets);
-
-        tickets.put(Status.TO_FIX, ticketRepo.countAllByProjectCreatorAndStatus(userEntity, Status.TO_FIX));
-        tickets.put(Status.IN_PROGRESS, ticketRepo.countAllByProjectCreatorAndStatus(userEntity, Status.IN_PROGRESS));
-        tickets.put(Status.COMPLETED, ticketRepo.countAllByProjectCreatorAndStatus(userEntity, Status.COMPLETED));
+        Map<Status, Long> ticketsStatusOverall = userServiceUtils.getTicketsStatusOverall(userEntity);
+        projectsInfoDTO.setTicketsStatusOverall(ticketsStatusOverall);
 
         return projectsInfoDTO;
     }
@@ -230,15 +219,15 @@ public class UserServiceImp implements UserService {
     @Override
     @NotNull
     public UserInfoDTO getUserInfo(String userId) {
-        UserEntity userEntity = serviceUtils.getUserEntity(userId);
+        UserEntity userEntity = dataExtractionUtils.getUserEntity(userId);
 
         long ticketsReported = ticketRepo.count(byReporter(userEntity));
         long ticketsReportedADayAgo = ticketRepo.count(byReporter(userEntity).and(reportedADayAgo()));
         long ticketsReportedAMonthAgo = ticketRepo.count(byReporter(userEntity).and(reportedAMonthAgo()));
 
-        long ticketsCompleted = ticketRepo.count(bySubscriber(userEntity).and(statusIsCompleted()));
-        long ticketsCompletedADayAgo = ticketRepo.count(bySubscriber(userEntity).and(statusIsCompleted()).and(reportedADayAgo()));
-        long ticketsCompletedAMonthAgo = ticketRepo.count(bySubscriber(userEntity).and(statusIsCompleted()).and(reportedAMonthAgo()));
+        long ticketsCompleted = ticketRepo.count(bySubscriber(userEntity).and(statusIs(Status.COMPLETED)));
+        long ticketsCompletedADayAgo = ticketRepo.count(bySubscriber(userEntity).and(statusIs(Status.COMPLETED)).and(reportedADayAgo()));
+        long ticketsCompletedAMonthAgo = ticketRepo.count(bySubscriber(userEntity).and(statusIs(Status.COMPLETED)).and(reportedAMonthAgo()));
 
         long ticketsSubscribedTo = ticketRepo.count(bySubscriber(userEntity));
         long ticketsSubscribedToADayAgo = ticketRepo.count(bySubscriber(userEntity).and(reportedADayAgo()));
@@ -257,29 +246,34 @@ public class UserServiceImp implements UserService {
         userInfoDTO.setTicketsSubscribedToADayAgo(ticketsSubscribedToADayAgo);
         userInfoDTO.setTicketsSubscribedToAMonthAgo(ticketsSubscribedToAMonthAgo);
 
+
+        Map<Severity, List<Long>> ticketsSeverityPerWeek = userServiceUtils.getTicketsSeverityPerWeek(userEntity);
+        userInfoDTO.setTicketsSeverityPerWeek(ticketsSeverityPerWeek);
+
+        Map<Severity, Long> tickets = userServiceUtils.getTicketsSeverityOverall(userEntity);
+        userInfoDTO.setTicketsSeverityOverall(tickets);
+
         return userInfoDTO;
     }
 
     public boolean isSubscribedToProject(@NotNull String userId, @NotNull String projectId) {
-        if (!projectRepo.existsByPublicId(projectId))
-            throw new ProjectServiceException(NO_RECORD_FOUND, projectId);
-        
-        UserEntity userEntity = serviceUtils.getUserEntity(userId);
+        if (!projectRepo.existsByPublicId(projectId)) throw new ProjectServiceException(NO_RECORD_FOUND, projectId);
+
+        UserEntity userEntity = dataExtractionUtils.getUserEntity(userId);
 
         return projectRepo.existsByPublicIdAndSubscribersIn(projectId, Set.of(userEntity));
     }
 
     public boolean isSubscribedToTicket(@NotNull String userId, @NotNull String ticketId) {
-        if (!ticketRepo.existsByPublicId(ticketId))
-            throw new TicketServiceException(NO_RECORD_FOUND, ticketId);
-        
-        UserEntity userEntity = serviceUtils.getUserEntity(userId);
+        if (!ticketRepo.existsByPublicId(ticketId)) throw new TicketServiceException(NO_RECORD_FOUND, ticketId);
+
+        UserEntity userEntity = dataExtractionUtils.getUserEntity(userId);
 
         return ticketRepo.existsByPublicIdAndSubscribersIn(ticketId, Set.of(userEntity));
     }
 
     public boolean isSuperAdmin(@NotNull String userId) {
-        return serviceUtils.getUserEntity(userId).getRoleEntity().getRole() == Role.ROLE_SUPER_ADMIN;
+        return dataExtractionUtils.getUserEntity(userId).getRoleEntity().getRole() == Role.ROLE_SUPER_ADMIN;
     }
 
 
@@ -287,7 +281,8 @@ public class UserServiceImp implements UserService {
     @Transactional
     public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
         UserEntity userEntity = userRepo.findByEmail(email)
-                .orElseThrow(() -> new UsernameNotFoundException(NO_RECORD_FOUND + " with : " + email));
+                .orElseThrow(() -> 
+                        new UsernameNotFoundException(NO_RECORD_FOUND + " with : " + email));
 
         UserDTO userDto = modelMapper.map(userEntity, UserDTO.class);
